Refactor section scroll listener
